Update named evaluation for HHVM
HHVM seems to cache file contents on `require` (and probably `include`)
so if names are reused, the initial version of the code will be required
again. To work around this, we always write to and require from a uniquely
named file, and then write the content to the named file for debugging.